[rename] MVI패턴을 위한 디렉터리 구조 세팅
