+++ app/src/main/java/com/example/sync_compose/MainActivity.kt
@@ -3,32 +3,18 @@ package com.example.sync_compose
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.*
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import androidx.navigation.NavHostController
-import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.composable
-import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
+import com.example.sync_compose.ui.navigation.BottomNavigationBar
+import com.example.sync_compose.ui.navigation.SetupNavGraph
+import com.example.sync_compose.ui.screen.home.HomeScreen
 import com.example.sync_compose.ui.theme.SynccomposeTheme
-import kotlinx.coroutines.launch
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -44,7 +30,6 @@ class MainActivity : ComponentActivity() {
 @Composable
 fun MainScreen() {
     val navController = rememberNavController()
-    val coroutineScope = rememberCoroutineScope()
 
     Scaffold(
         bottomBar = { BottomNavigationBar(navController = navController) }
@@ -52,109 +37,6 @@ fun MainScreen() {
         Box(modifier = Modifier.padding(innerPadding)) {
             SetupNavGraph(navController = navController)
         }
-    }
-}
-
-@Composable
-fun BottomNavigationBar(navController: NavHostController) {
-    val items = listOf(
-        BottomNavItem("home", "Home", R.drawable.ic_bnv_home),
-        BottomNavItem("community", "Community", R.drawable.ic_bnv_community),
-        BottomNavItem("chatting", "Chatting", R.drawable.ic_bnv_chat),
-        BottomNavItem("my", "My", R.drawable.ic_bnv_my)
-    )
-    Box(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clip(RoundedCornerShape(topStart = 24.dp, topEnd = 24.dp))
-            .background(Color.Transparent)
-            .border(1.dp, color = Color.Cyan, RoundedCornerShape(topStart = 24.dp, topEnd = 24.dp))
-    ) {
-        BottomNavigation(
-            modifier = Modifier.height(78.dp),
-            backgroundColor = Color.Black,
-            elevation = 20.dp
-        ) {
-            val navBackStackEntry by navController.currentBackStackEntryAsState()
-            val currentRoute = navBackStackEntry?.destination?.route
-            items.forEach { item ->
-                BottomNavigationItem(
-                    modifier = Modifier.padding(vertical = 16.dp),
-                    icon = {
-                        Icon(
-                            painterResource(id = item.icon),
-                            contentDescription = item.label
-                        )
-                    },
-                    label = { Text(item.label) },
-                    selected = currentRoute == item.route,
-                    selectedContentColor = Color.Blue,
-                    unselectedContentColor = Color.Red,
-                    onClick = {
-                        navController.navigate(item.route) {
-                            popUpTo(navController.graph.startDestinationId) {
-                                saveState = true
-                            }
-                            launchSingleTop = true
-                            restoreState = true
-                        }
-                    }
-                )
-            }
-        }
-    }
-
-}
-
-data class BottomNavItem(val route: String, val label: String, val icon: Int)
-
-@Composable
-fun SetupNavGraph(navController: NavHostController) {
-    NavHost(navController = navController, startDestination = "home") {
-        composable("home") { HomeScreen() }
-        composable("community") { CommunityScreen() }
-        composable("chatting") { ChattingScreen() }
-        composable("my") { MyScreen() }
-    }
-}
-
-@Composable
-fun HomeScreen() {
-    Box(
-        modifier = Modifier.fillMaxSize(),
-        contentAlignment = Alignment.Center
-    ) {
-        Text(text = "Home screen")
-    }
-}
-
-@Composable
-fun CommunityScreen() {
-    Box(
-        modifier = Modifier.fillMaxSize(),
-        contentAlignment = Alignment.Center
-    ) {
-        Text(text = "Community screen")
-    }
-}
-
-@Composable
-fun ChattingScreen() {
-    Box(
-        modifier = Modifier.fillMaxSize(),
-        contentAlignment = Alignment.Center
-    ) {
-        Text(text = "Chatting screen")
-    }
-}
-
-@Composable
-fun MyScreen() {
-    Box(
-        modifier = Modifier.fillMaxSize(),
-        contentAlignment = Alignment.Center
-    ) {
-        Text(text = "Settings screen")
     }
 }
