+++ app/src/main/java/com/example/sync_compose/ui/screen/community/CommunityIntent.kt
@@ -1,0 +1,6 @@
+package com.example.sync_compose.ui.screen.community
+
+sealed class CommunityIntent {
+    object LoadData : CommunityIntent()
+    data class ShowMessage(val message: String) : CommunityIntent()
+}
