+++ app/src/main/java/com/example/sync_compose/ui/screen/community/CommunityViewModel.kt
@@ -1,0 +1,6 @@
+package com.example.sync_compose.ui.screen.community
+
+import androidx.lifecycle.ViewModel
+
+class CommunityViewModel : ViewModel()  {
+}
