+++ app/src/main/java/com/example/sync_compose/ui/screen/home/HomeScreen.kt
@@ -1,0 +1,18 @@
+package com.example.sync_compose.ui.screen.home
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+
+@Composable
+fun HomeScreen() {
+    Box(
+        modifier = Modifier.fillMaxSize(),
+        contentAlignment = Alignment.Center
+    ) {
+        Text(text = "Home screen")
+    }
+}
